Artificial-Intelligence/Project/backend/requirements.txt
# Backend dependencies for Breast Tumor Diagnosis System
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0

# Traditional AI - Fuzzy Logic
scikit-fuzzy>=0.4.2

# Visualization
matplotlib>=3.7.0
opencv-python>=4.8.0

# Data handling
pandas>=2.0.0

# CORS
python-dotenv>=1.0.0

